Rename lineage.js to hydration.js
